# openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/music/search": {
      "get": {
        "operationId": "MusicController_search",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Music"
        ]
      }
    },
    "/music/prepare": {
      "post": {
        "operationId": "MusicController_prepare",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlaySongDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          }
        },
        "tags": [
          "Music"
        ]
      }
    },
    "/music/stream/{id}": {
      "get": {
        "operationId": "MusicController_streamSong",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "flac",
            "required": true,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Music"
        ]
      }
    },
...
    "/music/similar-tracks/{id}": {
      "get": {
        "operationId": "MusicController_getSimilarTracks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Music"
        ]
      }
    },
    "/music/check-flac/{id}": {
      "get": {
        "operationId": "MusicController_checkFlacAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Music"
        ]
      }
    },
...
    "/library/play/{id}": {
      "post": {
        "operationId": "LibraryController_recordPlay",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Library"
        ]
      }
    },
    "/playlists": {
      "get": {
        "operationId": "PlaylistController_getUserPlaylists",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Playlist"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      },
      "post": {
        "operationId": "PlaylistController_createPlaylist",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      }
    },
    "/playlists/{id}": {
      "get": {
        "operationId": "PlaylistController_getPlaylist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      },
      "put": {
        "operationId": "PlaylistController_updatePlaylist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlaylistDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      },
      "delete": {
        "operationId": "PlaylistController_deletePlaylist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Playlist"
        ]
      }
    },
    "/playlists/{id}/songs": {
      "post": {
        "operationId": "PlaylistController_addSongToPlaylist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSongToPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistSong"
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      }
    },
    "/playlists/{id}/songs/{songId}": {
      "delete": {
        "operationId": "PlaylistController_removeSongFromPlaylist",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "songId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Playlist"
        ]
      }
    },
    "/playlists/import/spotify": {
      "post": {
        "operationId": "PlaylistController_importFromSpotify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportPlaylistDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Playlist"
                }
              }
            }
          }
        },
        "tags": [
          "Playlist"
        ]
      }
    },
...
    "/profile/search": {
      "get": {
        "operationId": "ProfileController_searchUsers",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/{username}": {
      "get": {
        "operationId": "ProfileController_getProfile",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/{username}/activity": {
      "get": {
        "operationId": "ProfileController_getProfileActivity",
        "parameters": [
          {
            "name": "username",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profile"
        ]
      }
    },
    "/profile/me": {
      "put": {
        "operationId": "ProfileController_updateMyProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "Profile"
        ]
      }
    }
  },
  "info": {
    "title": "Your API",
    "description": "Your API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PlaySongDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "artistName": {
            "type": "string"
          },
          "albumName": {
            "type": "string"
          },
          "albumCover": {
            "type": "string"
          },
          "lastFMLink": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "mbid": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "artistName"
        ]
      },
      "PlaylistSong": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "playlist": {
            "$ref": "#/components/schemas/Playlist"
          },
          "playlistId": {
            "type": "string"
          },
          "song": {
            "type": "object"
          },
          "songId": {
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "addedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "playlist",
          "playlistId",
          "song",
          "songId",
          "position",
          "addedAt"
        ]
      },
      "Playlist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "source": {
            "enum": [
              "user",
              "spotify",
              "youtube"
            ],
            "type": "string"
          },
          "externalId": {
            "type": "string"
          },
          "coverImage": {
            "type": "string"
          },
          "isEditable": {
            "type": "boolean"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistSong"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "source",
          "externalId",
          "coverImage",
          "isEditable",
          "user",
          "userId",
          "songs",
          "createdAt",
          "updatedAt"
        ]
      },
      "SearchHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "query": {
            "type": "string"
          },
          "filter": {
            "type": "string"
          },
          "searchedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "query",
          "filter",
          "searchedAt"
        ]
      },
      "Artist": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "lastFMLink": {
            "type": "string"
          },
          "externalListens": {
            "type": "number"
          },
          "mbid": {
            "type": "number"
          },
          "songs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "songs",
          "createdAt",
          "updatedAt"
        ]
      },
      "Song": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "artistName": {
            "type": "string"
          },
          "albumName": {
            "type": "string"
          },
          "duration": {
            "type": "number"
          },
          "albumCover": {
            "type": "string"
          },
          "spotifyId": {
            "type": "string"
          },
          "youtubeId": {
            "type": "string"
          },
          "externalUrl": {
            "type": "string"
          },
          "hasFlac": {
            "type": "boolean"
          },
          "slskPath": {
            "type": "string"
          },
          "artistId": {
            "type": "number"
          },
          "lastFMLink": {
            "type": "string"
          },
          "externalListens": {
            "type": "number"
          },
          "mbid": {
            "type": "string"
          },
          "downloadedPath": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "playCount": {
            "type": "number"
          },
          "likeCount": {
            "type": "number"
          },
          "commentCount": {
            "type": "number"
          },
          "repostCount": {
            "type": "number"
          },
          "librarySongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibrarySong"
            }
          },
          "playlistSongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlaylistSong"
            }
          },
          "playHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayHistory"
            }
          },
          "artist": {
            "$ref": "#/components/schemas/Artist"
          },
          "relatedSongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Song"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "artistName",
          "albumName",
          "duration",
          "albumCover",
          "spotifyId",
          "youtubeId",
          "externalUrl",
          "hasFlac",
          "slskPath",
          "artistId",
          "lastFMLink",
          "externalListens",
          "mbid",
          "downloadedPath",
          "metadata",
          "playCount",
          "likeCount",
          "commentCount",
          "repostCount",
          "librarySongs",
          "playlistSongs",
          "playHistory",
          "artist",
          "relatedSongs",
          "createdAt",
          "updatedAt"
        ]
      },
      "PlayHistory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "songId": {
            "type": "string"
          },
          "playedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "song",
          "songId",
          "playedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string"
          },
          "isPrivate": {
            "type": "boolean"
          },
          "googleId": {
            "type": "string"
          },
          "isEmailConfirmed": {
            "type": "boolean"
          },
          "emailConfirmToken": {
            "type": "string"
          },
          "resetPasswordToken": {
            "type": "string"
          },
          "apiKey": {
            "type": "string"
          },
          "role": {
            "enum": [
              "user",
              "admin"
            ],
            "type": "string"
          },
          "subscriptionPlan": {
            "enum": [
              "free",
              "premium"
            ],
            "type": "string"
          },
          "subscriptionExpiresAt": {
            "format": "date-time",
            "type": "string"
          },
          "swagz": {
            "type": "number"
          },
          "playlists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Playlist"
            }
          },
          "librarySongs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LibrarySong"
            }
          },
          "searchHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SearchHistory"
            }
          },
          "playHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlayHistory"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "email",
          "password",
          "username",
          "bio",
          "avatarUrl",
          "isPrivate",
          "googleId",
          "isEmailConfirmed",
          "emailConfirmToken",
          "resetPasswordToken",
          "apiKey",
          "role",
          "subscriptionPlan",
          "subscriptionExpiresAt",
          "swagz",
          "playlists",
          "librarySongs",
          "searchHistory",
          "playHistory",
          "createdAt",
          "updatedAt"
        ]
      },
      "LibrarySong": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userId": {
            "type": "string"
          },
          "song": {
            "$ref": "#/components/schemas/Song"
          },
          "songId": {
            "type": "string"
          },
          "isLiked": {
            "type": "boolean"
          },
          "isDownloaded": {
            "type": "boolean"
          },
          "addedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "user",
          "userId",
          "song",
          "songId",
          "isLiked",
          "isDownloaded",
          "addedAt"
        ]
      },
...
      "CreatePlaylistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdatePlaylistDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "AddSongToPlaylistDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "artist": {
            "type": "string"
          },
          "album": {
            "type": "string"
          },
          "duration": {
            "type": "string"
          },
          "coverUrl": {
            "type": "string"
          }
        },
        "required": [
          "title",
          "artist"
        ]
      },
      "ImportPlaylistDto": {
        "type": "object",
        "properties": {
          "playlistUrl": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": [
          "playlistUrl"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri"
          },
          "isPrivate": {
            "type": "boolean"
          }
        }
      }
    }
  }
}